% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscanner-package.R
\docType{package}
\name{mapscanner-package}
\alias{mapscanner}
\alias{mapscanner-package}
\title{mapscanner: What the Package Does}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Authors:
\itemize{
  \item Michael D. Sumner \email{mdsumner@gmail.com}
}

}
\keyword{internal}
