% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_poly.R
\name{ms_aggregate_poly}
\alias{ms_aggregate_poly}
\title{Aggregate disparate polygons}
\usage{
ms_aggregate_poly(px, ...)
}
\arguments{
\item{px}{input polygons (assumed overlapping poly/mpolys in sf_df)}

\item{...}{unused}
}
\description{
Planar partition from disparate polygon inputs. Overlaps aggregate to `n`.
}
\details{
Input is a single simple features polygon data frame. No attribute data is considered.
}
\examples{
pts <- sf::st_sf(a = 1:3,
    geometry = sf::st_sfc(list(sf::st_point(cbind(0, 0)), sf::st_point(cbind(0, 1)), sf::st_point(cbind(1, 0)))))
overlapping_polys <- sf::st_buffer(pts, 0.75)

## decompose and count space-filling from overlapping polygons
x <- ms_aggregate_poly(overlapping_polys); plot(x)
#library(ggplot2)
#ggplot(x) + geom_sf() + facet_wrap(~n)
}
