% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{ms_get_map}
\alias{ms_get_map}
\title{ms_get_map}
\usage{
ms_get_map(bbox, max_tiles = 16L)
}
\arguments{
\item{bbox}{Either a string specifying the location, or a numeric bounding
box as a single vector of (xmin, ymin, xmax, ymax), or a 2-by-2 matrix with
columns of (min, max) and rows of (x, y), respectively.}

\item{max_tiles}{Maximum number of tiles to use to create map}
}
\value{
A map
}
\description{
Get a map image for a specified area or bounding box
}
