% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{ms_map_to_pdf}
\alias{ms_map_to_pdf}
\title{print map to a pdf file}
\usage{
ms_map_to_pdf(my_map, file)
}
\arguments{
\item{my_map}{A map produced with \link{ms_get_map}.}

\item{file}{Name of pdf file to print to (extension with be automatically
added).}
}
\value{
Nothing
}
\description{
print map to a pdf file
}
