% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify-maps.R
\name{ms_rectify_maps}
\alias{ms_rectify_maps}
\title{ms_rectify_maps}
\usage{
ms_rectify_maps(map_original, map_modified, type = "polygons",
  downsample = 10, quiet = FALSE)
}
\arguments{
\item{map_original}{File name of the original map without anything drawn over
it (either a `.pdf` or `.png`; extension will be ignored).}

\item{map_modified}{File name of the modified version with drawings (either a
`.pdf` or `.png`; extension will be ignored).}

\item{type}{Currently either "points", "polygons", or "hulls", where
"points" simply reduces each distinct object to a single, central point;
"polygons" identifies individual groups and returns the polygon representing
the outer boundary of each; and "hulls" constructs convex polygons around
each group.}

\item{downsample}{Factor by which to downsample polygons, noting that
polygons initially include every outer pixel of image, so can generally be
downsampled by at least an order or magnitude (`n = 10`). Higher values may
be used for higher-resolution images; lower values will generally only be
necessary for very low lower resolution images.}

\item{quiet}{If `FALSE`, display progress information on screen}
}
\value{
An \pkg{sf} object representing the drawn additions to map_modified.
}
\description{
Scan in two pdf or png maps, rectify them with `RNiftyReg`, and return the
modifications in `map_modified` as spatial objects in \pkg{sf} format.
}
\note{
Currently only return a single convex polygon surrounding all elements
added to `map_modified`.
}
\examples{
f_orig <- system.file ("extdata", "omaha.png", package = "mapscanner")
f_mod <- system.file ("extdata", "omaha-polygons.png", package = "mapscanner")
\dontrun{
xy_hull <- ms_rectify_maps (f_orig, f_mod, type = "hull")
xy_poly <- ms_rectify_maps (f_orig, f_mod, type = "polygons")
xy_pts <- ms_rectify_maps (f_orig, f_mod, type = "points")
}
# reduce file sizes to 1/4 to speed up these examples:
f_orig2 <- file.path (tempdir (), "omaha.png")
f_modified2 <- file.path (tempdir (), "omaha-polygons.png")
magick::image_read (f_orig) \%>\%
    magick::image_resize ("25\%") \%>\%
    magick::image_write (f_orig2)
magick::image_read (f_mod) \%>\%
    magick::image_resize ("25\%") \%>\%
    magick::image_write (f_modified2)

# then rectify those files:
xy_hull <- ms_rectify_maps (f_orig2, f_modified2, type = "hull")
xy_poly <- ms_rectify_maps (f_orig2, f_modified2, type = "polygons")
xy_pts <- ms_rectify_maps (f_orig2, f_modified2, type = "points")

}
