% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-maps.R
\name{scan_maps}
\alias{scan_maps}
\title{scan_maps}
\usage{
scan_maps(map_original, map_modified)
}
\arguments{
\item{map_original}{The original map without anything drawn over it.}

\item{map_modified}{The modified version with drawings}
}
\value{
An \pkg{RNiftyReg} object
}
\description{
Scan in two pdf maps and rectify them with `RNiftyReg`
}
