% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{set_mapbox_token}
\alias{set_mapbox_token}
\title{set_mapbox_token}
\usage{
set_mapbox_token(token)
}
\arguments{
\item{token}{Personal mapbox API token, obtained from
\url{https://docs.mapbox.com/api/#access-tokens-and-token-scopes}.}
}
\value{
`TRUE` if the token was able to be set; otherwise `FALSE`.
}
\description{
Set a mapbox token for use with the \link{ms_generate_map} function.
}
