% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscanner-package.R
\docType{package}
\name{mapscanner-package}
\alias{mapscanner}
\alias{mapscanner-package}
\title{mapscanner: Print Maps, Draw on Them, Scan Them Back In}
\description{
Print maps, draw on them, scan them back in, and convert to spatial
    objects.
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Authors:
\itemize{
  \item Michael D. Sumner \email{mdsumner@gmail.com}
}

}
\keyword{internal}
