% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscanner-package.R
\docType{package}
\name{mapscanner-package}
\alias{mapscanner}
\alias{mapscanner-package}
\title{mapscanner: Print Maps, Draw on Them, Scan Them Back in}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Enables preparation of maps to be printed and drawn on. Modified maps can then be scanned back in, and hand-drawn marks converted to spatial objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/mapscanner}
  \item Report bugs at \url{https://github.com/ropensci/mapscanner/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Authors:
\itemize{
  \item Michael D. Sumner \email{mdsumner@gmail.com}
}

Other contributors:
\itemize{
  \item Kelly Hondula (Kelly reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/330) [reviewer]
  \item Nicholas Potter (Nichola reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/330) [reviewer]
  \item Stanislaw Adaszewski (author of include concaveman-cpp code) [copyright holder]
}

}
\keyword{internal}
