% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{ms_generate_map}
\alias{ms_generate_map}
\title{ms_generate_map}
\usage{
ms_generate_map(bbox, max_tiles = 16L, mapname, raster_brick = NULL)
}
\arguments{
\item{bbox}{Either a string specifying the location, or a numeric bounding
box as a single vector of (xmin, ymin, xmax, ymax), or a 2-by-2 matrix with
columns of (min, max) and rows of (x, y), respectively.}

\item{max_tiles}{Maximum number of tiles to use to create map}

\item{mapname}{Name of map to be produced, optionally including full path.
Extension will be ignored.}

\item{raster_brick}{Instead of automatically downloading tiles within a given
`bbox`, a pre-downloaded `raster::rasterBrick` object may be submitted and
used to generate the `.pdf` and `.jpg` equivalents.}
}
\value{
Invisibly returns `TRUE` is both `.pdf` and `.jpg` maps successfully
created.
}
\description{
Generate a map image for a specified area or bounding box. Map is
automatically save in both `.pdf` and `.jpg` formats.
}
