% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify-maps.R
\name{ms_rectify_maps}
\alias{ms_rectify_maps}
\title{ms_rectify_maps}
\usage{
ms_rectify_maps(map_original, map_modified, type = "polygons",
  downsample = 10)
}
\arguments{
\item{map_original}{File name of the original map without anything drawn over
it (either a `.pdf` or `.jpg`; extension will be ignored).}

\item{map_modified}{File name of the modified version with drawings (either a
`.pdf` or `.jpg`; extension will be ignored).}

\item{type}{Currently either "points", "polygon", or "polygons", where
"points" simply returns the identified points that differ between the two
maps, "polygon" drawns a single convex hull around all points, and "polygons"
separates points into groups, returning polygons around each group.}

\item{downsample}{Factor by which to downsample polygons, noting that
polygons initially include every outer pixel of image, so can generally be
downsampled by at least an order or magnitude (`n = 10`). Higher values may
be used for higher-resolution images; lower values will generally only be
necessary for very low lower resolution images.}
}
\value{
An \pkg{sf} object representing the drawn additions to map_modified.
}
\description{
Scan in two pdf or jpg maps, rectify them with `RNiftyReg`, and return the
modifications in `map_modified` as spatial objects in \pkg{sf} format.
}
\note{
Currently only return a single convex polygon surrounding all elements
added to `map_modified`.
}
